Start removing dependency on Test::Output.
Currently many test blocks in t/Path.t require CPAN module Test::Output.
While this makes for very clean code, it's unsuitable for use in a module
distributed with the Perl 5 core.  Those tests are SKIPped if Test::Output is
not present, which is always the case when running the Perl 5 core test suite
and which is often the case with CPANtesters who prize a clean environment.

But if tests are skipped, important parts of File::Path's functionality are
never tested.  Given the range of Perl versions and operating systems on which
File::Path is expected to run, this is unacceptable.

This patch starts the process of removing the dependency on Test::Output by
reverting, for 5 unit tests not run by root, to the older, admittedly less
elegant $SIG{__WARN__} approach to capturing STDERR.