Test for unrecognized options.
In v2.10, we began to cause make_path and remove_tree to carp on unrecognized
options.  We originally wanted to have them croak, but ran into downstream
complaints, so we went with carp.  However, the tests for croaking were
commented out.

This patch replaces the commented-out tests with tests to detect warnings on
unrecognized options.