Add explicit tests for internal sub _error().
Most of the instances in which File::Path::_error() was hitherto exercised fell
into the blocks only run by root.  Hence, most of the branches and conditions
within that subroutine showed up as uncovered during coverage analysis, and it's
difficult for mere mortal users to get an idea of how the sub works.  (And
running Devel::Cover as root does not sound like a good idea.)

Add a block of tests which exercise all the nooks and crannies of
File::Path_error().